Cleaned up Javascript for most popular a bit
