+++ common/app/assets/javascripts/commonPlugins.js
@@ -1,16 +1,4 @@
-require(["reqwest"], function(reqwest){
-    reqwest({
-        url: guardian.page.coreNavigationUrl + '/most-popular/' + guardian.page.edition + '/' + guardian.page.section,
-        type: 'jsonp',
-        jsonpCallback: 'callback',
-        jsonpCallbackName: 'showMostPopular',
-        success: function(json) {
-            if (json.html) {
-                document.getElementById("most-popular").innerHTML = json.html;
-            }
-        }
-    })
-});
+require('mostPopular');
 
 //todo only load for percentage of users
 require(['http://cdn.optimizely.com/js/' + guardian.page.optimizelyId + '.js'], function(optimizely){})
